Seperated some more math related functions from single files
