+++ 100-js-algo/utils/Math/factorial.js
@@ -1,0 +1,11 @@
+function factorial(number) {
+    let result = 1;
+    while (number > 0) {
+        result *= number;
+        number--;
+    }
+
+    return result;
+}
+
+module.exports = { factorial };
